Hand Project
This Python project involves computer vision techniques for hand detection and classification of American Sign Language (ASL) gestures. It captures frames from a webcam, detects hands using computer vision methods, processes the region around the hands by resizing it, and predicts the ASL gesture using a pre-trained classifier. The results of the predictions are then visualized on the output image, displaying the detected hand and the predicted ASL gesture.